Extract all 'tf' types to a separate file Rename and move factory-types
