+++ src/serialize.ts
@@ -1,8 +1,9 @@
 import * as convert from './convert'
 import Formula from './formula'
 import Serializer from './serializer'
-import { ContentType, TFNamedNode, TFBlankNode } from './types'
+import { ContentType} from './types'
 import IndexedFormula from './store'
+import { TFBlankNode, TFNamedNode } from './tf-types'
 
 /**
  * Serialize to the appropriate format
