+++ src/utils/termValue.ts
@@ -1,4 +1,4 @@
-import { TFTerm } from '../types'
+import { TFTerm } from '../tf-types'
 
 /** Retrieve the value of a term, or self if already a string. */
 export function termValue (node: TFTerm | string): string {
