+++ src/variable.ts
@@ -1,7 +1,8 @@
 import ClassOrder from './class-order'
 import Node from './node-internal'
-import { TermType, TFVariable, VariableTermType } from './types'
+import { TermType, VariableTermType } from './types'
 import * as Uri from './uri'
+import { TFVariable } from './tf-types'
 
 /**
 * Variables are placeholders used in patterns to be matched.
